Optimize and refactor code
